requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
nltk>=3.6.0
emoji>=1.6.0
flask>=2.0.0
flask-cors>=5.0.0
gunicorn>=20.1.0
PyMuPDF>=1.18.0
python-docx>=0.8.11
pytesseract>=0.3.8
Pillow>=8.2.0
requests>=2.25.0
